project directory refactoring. added src and project_setup.py
