+++ README.md
@@ -69,7 +69,7 @@ Conda를 사용하지 않는 환경에서는 `pip`을 사용하여 Python 패키
 ## 💡 사용법
 
 ### 1. 설정 구성
-`config/config.yaml` 파일을 열어 데이터셋 경로(`root_dir`)와 훈련에 필요한 하이퍼파라미터를 수정하세요.
+`configs/config.yaml` 파일을 열어 데이터셋 경로(`root_dir`)와 훈련에 필요한 하이퍼파라미터를 수정하세요.
 
 ### 2. 탐색적 데이터 분석 (EDA)
 데이터의 특징을 파악하고 전처리 방향을 결정하기 위해 Jupyter Notebook을 실행합니다. 먼저, 프로젝트의 Conda 환경이 활성화되었는지 확인하세요.
@@ -82,7 +82,7 @@ jupyter lab notebooks/01_EDA.ipynb
 다음 명령어로 모델 훈련을 시작합니다. 훈련 로그는 logs/ 디렉토리에, 모델 가중치는 checkpoints/ 디렉토리에 저장됩니다.
 
 # (doc-classifier-env)
-python train.py --config config/config.yaml
+python train.py --config configs/config.yaml
 
 4. 추론 실행
 훈련된 모델을 사용하여 새로운 이미지에 대한 예측을 수행합니다.
